Move circuit_drawer_test.py to text_circuit_drawer_test.py
